Create separate module for UART functions
