+++ Makefile
@@ -21,12 +21,12 @@ SERIAL_PORT=/dev/ttyACM0
 #flags:
 #-j : only keep this section
 all:
-#	$(CC) -Wall -g2 -gstabs -O0 -fpack-struct -fshort-enums -ffunction-sections -fdata-sections -std=gnu99 -funsigned-char -funsigned-bitfields -mmcu=atmega328p -DF_CPU=16000000UL -MMD -MP -MF"main.d" -MT"main.d" -c -o main.o main.c
+	$(CC) $(CFLAGS) -c -o UARTComms.o UARTComms.c
 	$(CC) $(CFLAGS) -c -o main.o main.c
-	$(CC) -Wl,-Map,$(TARGET).map -mmcu=$(MCU) -o $(TARGET).elf  main.o  -lm
+	$(CC) -Wl,-Map,$(TARGET).map -mmcu=$(MCU) -o $(TARGET).elf  UARTComms.o main.o -lm
 	$(OBJCOPY) -j .text -j .data -O ihex $(TARGET).elf $(TARGET).hex
 	$(AVRSIZE) --format=$(FORMAT) --mcu=$(MCU) $(TARGET).elf
-
+	
 flash:
 	avrdude -p$(PARTNUM) -c$(PROG_TYPE) -b$(BAUD) -U$(MEM_OP_FLASH) -F -P$(SERIAL_PORT)
