+++ main.c
@@ -82,22 +82,16 @@
 #include <math.h>			//NB: Need to add m library in Library section
 
 
-//#include "debug.h"
+#include "UARTComms.h"
 
 //DEFINES: ------------------------
 #define BLINK_DELAY_MS 100
-//UART
-#define FOSC 16000000
-//#define BAUD 9600
-uint32_t BAUD = 115200; //9600
-//#define MYUBRR FOSC/16/BAUD-1   //Needed to define the UBRR (UART Baud rate) register
-uint32_t MYUBRR;
-
 
 
 //TIMER - DATA SAMPLING
 //NB: If SAMPLE_RATE Set to 0, use SECS_PER_SAMPLE for Timer configuration
 //	 For 1 sample/sec, use SAMPLE_RATE = 0; SECS_PER_SAMPLE = 1
+#define FOSC 16000000
 #define SAMPLE_RATE			1000 //1000   //(samples per sec)
 #define SECS_PER_SAMPLE		1
 #define PRESCALER			8  //64  //1024: Sample rate = 0; Others: USe Timer1_Calcs.xls table
@@ -141,13 +135,6 @@ uint32_t MYUBRR;
 
 
 //FUNCTION DECLARATIONS: --------------------------
-//UART:
-void ioinit(void);
-void USART_Transmit( int8_t data );
-void debug(char *str_data);
-static int uart_putchar(char c, FILE *stream);
-uint8_t uart_getchar(void);
-
 
 //Timer functions for data sampling (16 bit timer)
 void timer1_init(void);
@@ -190,8 +177,6 @@ int16_t getVal(int8_t msb,int8_t lsb);
 void doBaudTest(void);
 void freqResponse(void);
 
-
-static FILE usart_str = FDEV_SETUP_STREAM(uart_putchar, uart_getchar, _FDEV_SETUP_RW);
 //-------------------------------------------------
 
 //Define flag ('s': send substitute; 'n': send unchanged)
@@ -438,117 +423,6 @@ void simulation(void)
 
 }//simulation
 
-
-void ioinit()
-{
-	//Set up the UART Registers:
-	//Override define definition to avoid rounding errors (See datasheet)
-
-
-
-	if(BAUD == 4800)
-	{
-		MYUBRR = 207;
-	}
-	if(BAUD == 9600)
-	{
-		MYUBRR = 103;
-	}
-	if(BAUD == 14400)
-	{
-		MYUBRR = 68;
-	}
-	if(BAUD == 19200)
-	{
-		MYUBRR = 51;
-	}
-	if(BAUD == 38400)
-	{
-		MYUBRR = 25;
-	}
-	if(BAUD == 57600)
-	{
-		MYUBRR = 16;
-	}
-	if(BAUD == 115200)
-	{
-		MYUBRR = 8;
-	}
-	if(BAUD == 230400)
-	{
-		MYUBRR = 3;
-	}
-
-
-	UBRR0H = MYUBRR >> 8;  //(Essentially, sets high bits to zero)
-	UBRR0L =  MYUBRR;
-	UCSR0B = (1<<RXEN0)|(1<<TXEN0); //Sets the required bits to enable the UART Tx and Rx
-
-	UCSR0A = (1<<UDRE0); //Set flag to empty data register - then ready to accept new data
-
-	UCSR0C =  (1 << UCSZ01) | (1 << UCSZ00);    // Set character size to 8 bit,
-												//NB:1 stop bit (USBS0 bit)  0 by default
-												//Parity bit (UPM0/OPM1) zero by default
-
-	//For standard comms (e.g. settings in Putty)
-	//Baud rate:  9600
-	//Bits: 8
-	//Stop: 1
-	//Parity: None
-
-	stdout = &usart_str; //Required for printf() init
-	stdin = &usart_str;  //Required for scanf() init
-
-}
-
-void USART_Transmit(int8_t data )
-{
-    /* Wait for empty transmit buffer */
-    while ( !( UCSR0A & (1<<UDRE0)) )
-    ;
-    /* Put data into buffer, sends the data */
-    UDR0 = data;
-}
-
-void debug(char *str_data)   //Send string to PC console
-{
-    while(*str_data)
-    {
-    	USART_Transmit(*str_data);
-        str_data++;
-    }
-}
-
-static int uart_putchar(char c, FILE *stream) {
-
-	//REGISTER DESCRIPTIONS
-	//UDR0 is the data register for current data (either receive or transmit) - see Datasheet
-	//If written here, it will then be transferred to transfer Data register TXB
-	//UCSR0A - Control and data register
-	//UDRE0 - flag (bit 5) of UCSR0A - if flagged to 1, then it means the data register isready for the
-	//next value
-
-	if (c == '\n')
-        uart_putchar('\r', stream);
-    loop_until_bit_is_set(UCSR0A, UDRE0);
-
-
-    UDR0 = c;
-
-    return 0;
-}
-
-uint8_t uart_getchar(void)
-{
-    //REGISTER DESCRIPTIONS:
-	//UDR0 is the data register for current data (either receive or transmit) - see Datasheet
-	//UCSR0A - Control and data register
-	// RXC0 flag (bit 7) is the flag to indicate that the data has been read (and buffer is now empty)
-
-	while( !(UCSR0A & (1<<RXC0)) );
-
-    return(UDR0);
-}
